[REFACTOR] 디렉토리 분리 (#60)
